Add script filters for test support files
