+++ src/test/groovy/bugs/scriptForGroovy1567.groovy
@@ -16,15 +16,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-a = 1
-evaluate("a = 3")
-assert a == 3
 
-println("Done and now set a to ${a}")
-
-
-println("About to call another script")
-
-evaluate(new File("src/test/groovy/script/HelloWorld.groovy"))
-
-println("Done")
+t = new Thread({ println "Groovy" })
+t.start()
+t.join()
