+++ src/test/groovy/script/scriptHelloWorld.groovy
