+++ src/test/groovy/script/scriptHelloWorld2.groovy
