+++ src/test/groovy/script/scriptWithClass.groovy
