+++ src/test/groovy/script/scriptWithClosure.groovy
