+++ src/test/groovy/script/scriptWithPackageStatement.groovy
